+++ tests/DeEssPresetTests.cpp
@@ -113,6 +113,7 @@ namespace
             { BinaryData::chestResonanceTamer_json, BinaryData::chestResonanceTamer_jsonSize },
             { BinaryData::fastRecoveryDemo_json, BinaryData::fastRecoveryDemo_jsonSize },
             { BinaryData::listenCheck_json, BinaryData::listenCheck_jsonSize },
+            { BinaryData::analogWarmthLift_json, BinaryData::analogWarmthLift_jsonSize },
         };
     }